Refactor: Remove unused user and owner model classes; add EVOwner model and related classes
