+++ Controllers/EVOwnersController.cs
@@ -266,24 +266,4 @@ namespace SparkPoint_Server.Controllers
             }
         }
     }
-
-    public class EVOwnerRegisterModel
-    {
-        public string Username { get; set; }
-        public string Email { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Password { get; set; }
-        public string NIC { get; set; }
-        public string Phone { get; set; }
-    }
-
-    public class EVOwnerUpdateModel
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-        public string Phone { get; set; }
-        public string Password { get; set; }
-    }
 }
