+++ Controllers/UsersController.cs
@@ -307,30 +307,4 @@ namespace SparkPoint_Server.Controllers
             }
         }
     }
-
-    public class UserUpdateModel
-    {
-        public string Username { get; set; }
-        public string Email { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Password { get; set; }
-    }
-
-    public class UserListFilterModel
-    {
-        public int? RoleId { get; set; }
-        public bool? IsActive { get; set; }
-        public string SearchTerm { get; set; }
-    }
-
-    public class CreateStationUserModel
-    {
-        public string Username { get; set; }
-        public string Email { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Password { get; set; }
-        public string ChargingStationId { get; set; }
-    }
 }
